refactor: all in one
