+++ .gitignore
@@ -1,6 +1,5 @@
 /target
 /script
-docker-compose.prod.yaml
 Makefile.*
 config.toml
 *.json
