+++ Cargo.toml
@@ -1,29 +1,15 @@
 [package]
 name = "janus"
-version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "discv"
-
-[[bin]]
-name = "ping"
-
-[[bin]]
-name = "server"
-
-[[bin]]
-name = "status"
-
 
 [dependencies]
 tokio = { version = "1.52", features = ["full"] }
 discv4 = { git = "https://github.com/rllola/discv4" }
 secp256k1 = { version = "0.29", features = ["recovery", "global-context", "rand-std"] }
 rand = "0.10"
-sqlx = { version = "0.8", default-features = false, features = ["runtime-tokio", "tls-rustls", "postgres", "json"] }
+sqlx = { version = "0.8", default-features = false, features = ["runtime-tokio", "tls-rustls", "postgres", "json", "migrate", "macros"] }
 hex = { version = "0.4.3", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.9"
@@ -39,7 +25,6 @@ byteorder = "1.5"
 hmac-sha256 = "1.1"
 ctr = "0.10"
 k256 = { version = "0.14", features = ["ecdh", "ecdsa-core", "arithmetic"] }
-postgres = { version = "0.19", features = ["with-serde_json-1"] }
 
 [dev-dependencies]
 devp2p = { git = "https://github.com/rllola/devp2p" }
