+++ Dockerfile
@@ -1,26 +1,28 @@
-##### BUILDER #####
-FROM rust:latest as builder
+# ── Builder ───────────────────────────────────────────────────────────────────
+FROM rust:slim AS builder
 
-WORKDIR /usr/src/prototype
+WORKDIR /usr/src/janus
+
 COPY . .
 RUN cargo install --path .
 
-##### RUNNER #####
-FROM debian:bookworm
+# ── Runtime ───────────────────────────────────────────────────────────────────
+FROM debian:trixie-slim
 
 LABEL author="Lola Rigaut-Luczak <me@laflemme.lol>"
 LABEL description="Programs that map nodes network (Ethereum and Ethereum like)."
 
-COPY --from=builder /usr/local/cargo/bin/discv /usr/local/bin/discv
-COPY --from=builder /usr/local/cargo/bin/ping /usr/local/bin/ping
-COPY --from=builder /usr/local/cargo/bin/server /usr/local/bin/server
-COPY --from=builder /usr/local/cargo/bin/status /usr/local/bin/status
+COPY --from=builder /usr/local/cargo/bin/janus /usr/local/bin/janus
 
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    ca-certificates \
+    && rm -rf /var/lib/apt/lists/*
 
-RUN apt-get update && rm -rf /var/lib/apt/lists/*
-#RUN apt-get install -y libc6
+# TCP/UDP port server
+EXPOSE 30303/tcp
+EXPOSE 30303/udp
 
-# default env
-ENV RUST_LOG "info"
+# Default log is info
+ENV RUST_LOG="janus=info"
 
-CMD discv
+ENTRYPOINT ["janus"]
